50Plots/requirements.txt
mysql-connector-python
argparse
black
pandas